Edit package name again
Romanesco seems to be a better fit after all, instead of RomanescoPatterns. Because the package name is being used throughout MODX, it makes things look much cleaner if it's just Romanesco. The internal reference for the git repository will remain romanesco-patterns though, because it's a good thing that this reveals some information about its contents already.