Show chunk based previews in slider
The best way to show an example of a chunk based pattern, is probably to enter a chunk call in an HTML Code field and show the result. Although it's just an example, it could be useful to see this code and use it as a starting point for your own implementation of the pattern.

So in cases where the example is displayed through a chunk, this chunk code gets displayed first now in the second window of a slider. This makes it feel more natural; it shows that that code is "behind" the given example. After being used for the code example, the code gets parsed into the visual example.

And if there's no code example available, no slider will be initialized. So it's safe to use with other methods of creating examples, like using CBs. A chunk call will not work in all scenarios, unfortunately.