Load referring patterns based on assigned_templates setting
Initially, every referring element query was controlled by a checkbox in CB.
This proved to be very had to maintain, so instead it's just the assigned templates that's being handled this way now.

So if the Assigned Templates setting is not checked in CB, all the other elements will be displayed (if set).