Clean up pattern layout elements
The front-end library patterns will be loaded from a single template now, with a few variations for specific cases. Caching remains tricky when using chunks inside layout templates, but I found a workaround by setting the template to a placeholder first.

I also improved the preview slider by checking the label of the code field in the cbGetFieldContent call. It was looking at all the code fields before, and assigned an idx value to every field it found. This was a bit dodgy.. If you would add another code field anywhere in the layout, the idx of that code field would be reference by the next layout for the preview.

In other words: you could only use 1 Code (Display) CB in each layout. That's not true actually.. Putting the code field inside a nested layout didn't affect the idx of fields inside the root of the main layout, but that is probably an inconsistency (or a bug) in ContentBlocks itself. Meaning it will probably break at some point. By having the extra check for the label setting, it assigns the idx values only to the code fields with that label. The only drawback there is that it didn't accept lexicon keys, so the label has to be hardcoded in the CB setting. I chose to use MODX in this case.