Add ability to filter included CB list
I also tried to work my way around a MODX quirk that kept on displaying the items of the last snippet call in all the includedBosons lists. I couldn't solve it with placeholders this time, as it has something to do with how MODX traverses through the snippet calls. This is expected behavior and won't be addressed until MODX 3.0.

It works for now, but only if you place the includedBosons snippet call inside a ContentBlock layout. Moving it to the layout template (its preferred location) causes the duplicated items issue to emerge.