Replace example slider with popup
It just wasn't working out that well with the slider. Styling-wise and functionality-wise. I do want to keep the option though to see the underlying chunk call, so the content is moved to a little popup now.

Beware that every pattern layout must contain a Code (Display) CB with a label title of 'MODX'. Otherwise, the offset for the cbGetFieldContent will be set incorrectly!