Refactor library preview and MODX code example
Get rid of that impossible render_example construction based on a cbGetFieldContent call with a fieldSettingsFilter based on the IDX of the layout, which was automatically deducted by one, but turned out to be no longer accurate if not all the fields had a Code (Display) CB, and was depending on the example label being explicitly set to MODX, which could not be replaced by a lexicon string, etc. etc.

Horrible.

The :tag modifier saved me. Was able to make that work with a custom snippet (cbRenderCodeField).