Refactor Button CB
The chunk call inside the CB template field couldn't handle the addition of the icon button anymore. To prevent ending up with a really complicated and fragile collection of If statements, I created a wrapper chunk to handle the output modifying and tpl selection.

The buttonOuter chunk still has one caveat inside though: the mosquito technique doesn't work for some reason on the If call that checks whether an icon was selected (and loads buttonHrefBasic or buttonHrefIcon correspondingly). This means that both chunks will be rendered in the background, but only 1 will be displayed. It's probably not a massive performance hit, but it is unnecessarily rendering something now which should always be avoided if possible.

**NB!** The button settings were also refactored in CB, resulting in 2 new settings chunks for button options and icon options. This leads to a breaking change for all existing buttons with button_type='basic' and button_block='1' (for full width buttons). Both settings are under button_options now, which is a multiselect (checkboxes).