Add patterns for displaying header vertically
Usually the main navigation sits on top of the main content area, but an alternative layout pattern is to move the main menu to the left (or right) side of the content.

To do still: fix the header in place (so it's always visible when scrolling through content) and correct the margin / padding issues that arose from the move.