Fix small margin difference of toolbar container
This was caused by the container being a .grid.container. It's officially supported in SUI, but it led to this minor difference in left and right margin on smallest screen size. The margins were set in em, where regular containers are using rem.

Adding an extra wrapper div for the container aligned everything again.