Add outputAsTpl option to modifiedIf
When set to 1, this takes the then/else value as chunk name and renders the chunk after the If evaluation is done. You can also use the mosquito technique for this, but in complex scenarios (with a lot of additional properties for example) this sometimes fails or leads to caching quirks.