Add main navigation with dropdown submenus
The subs are arranged in one big popup, not the default drawers + fly-outs. This makes it easier to orientate yourself with the site structure and saves you a few (often annoying) interactions with the fly-outs as well.

In order for the 2nd level items to show their subs in the dropdown, they need to have `rel="category"` attached as link attribute.