Define responsive img scale factor based on layout type
This only works partially. It doesn't know what the exact column count of the layout is on a given breakpoint, because the layout settings for sidebar width and responsive options are not yet coming through.

Bug report has been filed and acknowledged by ModMore.

Currently, it loads all mobile and tablet images at 100vw (viewport width).