Add page number to unique_idx placeholder in Overviews
Although the pagination is handled by pdoTools now, the snippet for loading the results is still getResources. And getResources starts with an ID of 1 again when going to the next page in the pagination menu, meaning the unique ID of each item is no longer unique. This lead to incorrect content being loaded in the items when refreshing a paginated URL.