Add suffix with version number to main CSS and JS assets
This is to force a refresh after an update. Version number is stored in a system setting. Independent of any real package release numbers, because it would be hard to tell which update should trigger a version bump. Now it's done after every styling update coming from the Configuration settings and it can be updated manually as well, if needed.