Write theme.variables output directly to file
The method of fetching the contents of the theme.variables resource generated by MODX proved to be too unreliable. The content wasn't always up-to-date by the time it got fetched and sometimes it even returned a 502. Maybe because it was still being rendered and not yet cached.

This update relies on a custom project theme to be present in the SUI src/themes folder. This has a number of advantages:

- Better theming when using multiple contexts
- No need to inject any paths in the site folder
- No need for the local.less file either anymore
- Avoid having to write into the Romanesco theme folder

You need to add this project theme manually before attempting to build SUI for an existing project. You can copy/paste it from romanesco-soil.