Add option to generate critical CSS in sequence
By default, the processes run in parallel because the command is disowned (meaning PHP forgets about it and moves on). This becomes problematic when running multiple processes (> 10) on your server.

Running in sequence is of course much slower, but this at least prevents the server from running out of resources when using the snippet to generate a large batch of files.