Make sure unique_idx placeholder in overviews is really unique
For example: when loading the footer through another resource, this resource could generate the same uid as the containing resource. That in turn could result in previous content leaking over into footer overviews.

The added defense mechanism here is the layout title, which can be manually altered in case of problems. Those would be real edge cases though.