Scroll back to top after using pagination in overviews
Because the AJAX pagination doesn't refresh the page, you might see an empty gap if the content of the next page is shorter. Now, it scrolls back to the top of the overview to prevent this from occurring.

This can be disabled with a TV, in case multiple overviews are used on the same page.