Don't display globally excluded resources in Overviews either
This also improves how the list of excluded resources is processed. The user doesn't need to prepend the IDs with a dash in Configuration settings. Instead, the new beforeEach snippet will prefix the resource IDs and then set a global placeholder in the head chunk, for use by all navigation elements.