Enable resizeObserver for Swiper instances
This ensures that items receive the correct width.

Sometimes the slider only snapped into its correct dimensions after a resize of the browser window. This may have to do with the fact that CSS is loaded asynchronously now, when critical CSS is enabled. Maybe the slider data gets calculated before the CSS is fully loaded.