Don't remove rows in mobile reversed grids
This breaks nested layouts with multiple rows.

I can't retrieve when it was introduced, but the reversed order is affected by the rows. So I guess there's some scenario where the desired column doesn't end up on top.

This should be handled by CSS or JS though, as it is a responsive problem.