Rename cols field to content_width in 2 column fluid overview
In other overviews the cols field is used to indicate column count, but here it was defining the width of the content column.

That made it a bit ambiguous, because the responsive image pattern counts on cols to be the number of columns, not the width. On top of that, the cols where passed along as text value, where a number value was needed for calculating the img sizes.

This meant that all image sizes were broken in 2 column fluid overviews. To avoid further issues and confusion, I decided to change the name of the field setting to a more logical content_width.

This breaks backward compatibility, because old cols values other than 8 (the default) will need to be changed manually after update.