Turn external links TV into MIGXdb grid
It used to be a selector for weblinks, but that was hardly adding any useful functionality.

Now, it's a grid with the option to add a number, which you can refer to in the content. Good for referencing source material in scientific articles.

I don't think the old selector has ever been used, but nevertheless: it breaks backwards compatibility.