Disable Jquery fallback
And only load from local source in dev mode.

Although the fallback was not processed when the CDN file was fetched, this workaround still needed to download an extra file. All other efforts to provide a workaround in combination with defer have failed, so I'm reverting to not having a fallback.

I guess the only way to make this work is to bundle the JS assets using ES6 modules or something like Rollup or Webpack.