Improve controls for inlining SVGs
There is now a configuration setting (svg_inline) which decides if the SVG should be placed in an <img> tag, or inline in an <svg> tag (with svgSanitize).

This option is overruled by the lazy_load setting of the element containing the SVG(s), because lazy loading the requests still seems like a more effective optimization than inline display.

Needs the latest version (1.1.0) of svgSanitize installed!