Add column classes to nested grid containers
If a nested grid contains multiple columns, these columns are arranged according to their size, combined with the 'equal width' classes on the grid container.

This works well in most cases, but some grids don't scale down nicely on tablet / computer breakpoints because the parent doesn't know how many columns it should count on.

This addition sets these classes on the parent by counting the number of columns being parsed.

Only applies to nested grids containing a single row, as different column counts can be applied to multiple rows in the same grid.