Insert critical CSS inline in head
It doesn't seem to affect overall page load speed that much, but it does shorten the time to interaction a bit.

Reasons for not doing it like this in the first place included repeated page loads, where the HTML would include the critical CSS again, as opposed to the external CSS file already being cached.

Anyway, let's roll with it and test speeds across the board.