Fix Tabs not working inside CTAs
Apparently, the unique_idx placeholder is not always rendered together with the CTA content. This would lead to incorrect combinations of tab / segment identifiers.

To fix, the placeholder is now set explicitly with the uid of the CTA block in the host resource.

This is a partial fix, which will fail again if multiple tab CBs are added inside a single CTA (because single unique_idx).