Link to authors via their team page ID
BREAKING CHANGE

Authors are no longer linked to a MODX user, but directly tied to a team member resource.

This is more appropriate for presentation purposes and eliminates the need to use the (clumsy) extended user settings to link each user to the corresponding team page.

This also simplifies templating of all Publication parts, at the expense of having to create a team page for each author.